Custom Emojis
I read through the code in #31 and @rhysd's comments about what should
be returned. I then read through the MDAST nodes and there's an Html
node, so I used that for the custom emoji as well as changed the aria
accessibility wrapping to use that, making it simpler.

Added test cases and tweaked the accessibility ones.